Combined simple.json and geoserver scripts - will need to change simple.json path to a test path in future. Also allow update_database.sh to be run remotely
